Detect if "Draw over other apps" permission is missing and fixes it.
This permission should be granted automatically when installed from the Play Store.
But some privacy related apps can automatically disable this permission.
The fix consists in asking the System settings if the app is allowed to "draw over others".
If the permission is not allowed, the user is redirect to the corresponding settings screen.
This alos fixes #147